+++ satella/coding/decorators/decorators.py
@@ -116,10 +116,9 @@ def wraps(cls_to_wrap: tp.Type) -> tp.Callable[[tp.Type], tp.Type]:
             cls.__annotations__ = cls_to_wrap.__annotations__
         try:
             sig = inspect.signature(cls_to_wrap)
-        except (TypeError, ValueError):
-            pass
-        else:
             cls.__signature__ = sig
+        except (TypeError, ValueError, RecursionError):
+            pass
         return cls
 
     return outer