+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # v2.11.35
 
 * added `ListWrapperIterator`
+* fixed a bug where `wraps` would sometimes fail with RecursionError
+
+